requirements.txt
annotated-types==0.7.0
anyio==3.7.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
Flask==2.3.3
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
numpy==2.3.1
openai==1.96.1
packaging==25.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.31.0
sniffio==1.3.1
soundfile==0.13.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
deepgram-sdk
firebase-admin>=6.0.0
pyaudio>=0.2.11